board-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.1
minio==7.2.0
pdf2image==1.17.0
pytesseract==0.3.10
pypdf==3.17.1
python-docx==1.1.0
pymilvus==2.3.4
marshmallow==3.12.2
environs==9.5.0
email-validator==2.1.0
celery>=5.2.7
flower>=1.2.0
redis>=4.4.0 